create moment template reminder hook
